Refactor: Added wrapper to the Getourapp
