+++ public/icons/index.ts
@@ -1,4 +1,3 @@
-// public/icons/index.ts
 import GoogleDownloadButton from "./google-bownload-button.svg";
 import AppleDownloadButton from "./apple-download-button.svg";
 import Twitter from "./twiiter.svg";
@@ -10,6 +9,9 @@ import Telephone from "./telephone.svg";
 import Mail from "./mail.svg";
 import Apple from "./apple.svg";
 import Google from "./google.svg";
+import Ellipse1 from "./ellipse-1.svg";
+import Ellipse2 from "./ellipse-2.svg";
+import Ellipse3 from "./ellipse-3.svg";
 
 export {
   GoogleDownloadButton,
@@ -23,4 +25,7 @@ export {
   Mail,
   Apple,
   Google,
+  Ellipse1,
+  Ellipse2,
+  Ellipse3,
 };
