+++ public/images/index.ts
@@ -1,0 +1,4 @@
+import LeftPhone from "./left-phone.png";
+import RightPhone from "./right-phone.png";
+
+export { LeftPhone, RightPhone };
